% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{gh_neighbor}
\alias{gh_neighbor}
\title{Find neighbor of a geohash in the requested direction}
\usage{
gh_neighbor(geohash, direction)
}
\description{
Find neighbor of a geohash in the requested direction
}
